+++ demos/CineCue/app/helpers/movies_helper.rb
@@ -1,0 +1,2 @@
+module MoviesHelper
+end