+++ demos/Flag Quiz/tailwind.config.js
@@ -1,0 +1,27 @@
+/** @type {import('tailwindcss').Config} */
+module.exports = {
+  content: [
+    "./app/**/*.{js,ts,jsx,tsx,mdx}",
+    "./pages/**/*.{js,ts,jsx,tsx,mdx}",
+    "./components/**/*.{js,ts,jsx,tsx,mdx}",
+ 
+    // Or if using `src` directory:
+    "./src/**/*.{js,ts,jsx,tsx,mdx}",
+  ],
+  
+  theme: {
+    extend: {
+      animation: {
+        spinSlow: 'spin 10s linear infinite',
+      },
+      keyframes: {
+        slowRotate: {
+          '0%': { transform: 'rotate(-180deg)' },
+          '100%': { transform: 'rotate(0deg)' },
+        },
+      },
+    },
+  },
+  plugins: [],
+}
+